feat:前端頁面調整

+++ src/main/java/com/timshuns/service/BlogService.java
@@ -7,7 +7,7 @@ public interface BlogService {
 
   Blog getBlog(Long id);
 
-  Page<Blog> getBlogs(long currentPage,String title,int typeId,int published);
+  Page<Blog> getBlogs(long currentPage,String title,long typeId,int published);
 
   boolean saveBlog(Blog blog);
